🔧 Change the spell option to use Aspell
## Changes Description
After reaching a certain size on my configuration file, `flyspell` was
not an option anymore. I caused to many hangs, and the computer
processor was spiking all the time. The default alternative that Doom
offers has a fantastic speed, but it only works well with Aspell.

This commit changes the spell check module to use Aspell.