🐛 Fix blank line at the end of snippets
## Changes Description
Due to the nature of snippets, we can't have an empty line at the end of
the snippet. If a snippet require an empty line, it can add one with a
simple '\n' character.

Previously, I had a `when-let` on the block that would decide about
stripping empty lines, and the very first variable definition was
setting the var to `nil` 🤦.

This commit replaces the `when-let` by a simple `let` so the proper
code can run.